Fixed 'check_existence' Methods in Request Category in 'fog-azure-rm'
Summary:
The check_existence methods had been written to throw an exception if the resource group in which the existence of resource was being checked did not exist itself.
The methods have been changed to return false even if the resource group does not exist itself.

Test Plan: N/A

Subscribers: omair.sajid

Differential Revision: http://phabricator.confiz.com/D10961