Updated List Method for Network Security Groups (#410)
Summary: The list method for network security groups was using the 'list_as_lazy' method which was not handling pagination by default. It has been replaced with the 'list' method that returns all the results even if they are divided into pages. The unit tests and mocks have been updated accordingly as well.