Added Null Check for Path Params Attribute in Azure Exception
Summary: When printing READ and WRITE limits for Azure API requests, an exception occurs when trying to access a key inside a hash that is null. A check has been added to first ensure that the hash is not null before extracting the Subscription ID from it.

Test Plan: N/A

Subscribers: omair.sajid

Differential Revision: https://phabricator.confiz.com/D13145